refactor(rnode): keep bearer status focused
